ENG-6129 ensure that screen name shows in react native and adv react native versions, move common view items to ViewIdentifiers
