+++ NeuroID/src/advancedDeviceReactNativeLib/java/com/neuroid/tracker/callbacks/NIDActivityCallbacks.kt
@@ -20,8 +20,15 @@ class NIDActivityCallbacks() : ActivityCallbacks() {
         val existActivity = listActivities.contains(currentActivityName)
 
         NIDServiceTracker.screenActivityName = currentActivityName
-        NIDServiceTracker.screenFragName = ""
-        NIDServiceTracker.screenName = "AppInit"
+        if (NIDServiceTracker.firstScreenName.isNullOrEmpty()) {
+            NIDServiceTracker.firstScreenName = currentActivityName
+        }
+        if (NIDServiceTracker.screenFragName.isNullOrEmpty()) {
+            NIDServiceTracker.screenFragName = ""
+        }
+        if (NIDServiceTracker.screenName.isNullOrEmpty()) {
+            NIDServiceTracker.screenName = "AppInit"
+        }
 
         val changedOrientation = auxOrientation != orientation
         wasChanged = changedOrientation
