+++ NeuroID/src/reactNativeLib/java/com/neuroid/tracker/events/NIDIdentifyAllViews.kt
@@ -1,44 +1,35 @@
 package com.neuroid.tracker.events
 
-import android.os.Build
 import android.view.View
 import android.view.ViewGroup
-import android.text.TextWatcher
 import android.widget.*
 import android.widget.RadioGroup
 import androidx.core.view.children
 import androidx.core.view.forEach
-import androidx.annotation.RequiresApi
 import com.facebook.react.views.image.ReactImageView
 import com.facebook.react.views.text.ReactTextView
-import com.neuroid.tracker.callbacks.NIDLongPressContextMenuCallbacks
-import com.neuroid.tracker.callbacks.NIDTextContextMenuCallbacks
 import com.facebook.react.views.textinput.ReactEditText
 import com.facebook.react.views.view.ReactViewGroup
 import com.neuroid.tracker.extensions.getSHA256withSalt
 import com.neuroid.tracker.callbacks.NIDSensorHelper
 import com.neuroid.tracker.models.NIDEventModel
-import com.neuroid.tracker.models.NIDSensorModel
 import com.neuroid.tracker.service.NIDServiceTracker
-import com.neuroid.tracker.storage.getDataStoreInstance
-import com.neuroid.tracker.utils.NIDTextWatcher
 import com.neuroid.tracker.extensions.getIdOrTag
 import com.neuroid.tracker.extensions.getParents
 import org.json.JSONArray
 import org.json.JSONObject
 import com.neuroid.tracker.storage.NIDDataStoreManager
-import com.neuroid.tracker.utils.NIDLog
 import com.neuroid.tracker.utils.NIDLogWrapper
 
 fun identifyAllViews(
     viewParent: ViewGroup,
     guid: String,
-    registerTarget: Boolean,
-    registerListeners: Boolean,
     logger: NIDLogWrapper,
     storeManager: NIDDataStoreManager,
+    registerTarget: Boolean,
+    registerListeners: Boolean,
     activityOrFragment: String = "",
-    parent: String = "",
+    parent: String = ""
 ) {
     logger.d("NIDDebug identifyAllViews", "viewParent: ${viewParent.getIdOrTag()}")
 
@@ -65,10 +56,10 @@ fun identifyAllViews(
                 identifyAllViews(
                     it,
                     guid,
-                    registerTarget,
-                    registerListeners,
                     logger,
                     storeManager,
+                    registerTarget,
+                    registerListeners,
                     activityOrFragment,
                     parent
                 )
@@ -141,10 +132,10 @@ fun identifyView(
         is ViewGroup -> identifyAllViews(
             view,
             guid,
-            registerTarget,
-            registerListeners,
             logger,
             storeManager,
+            registerTarget,
+            registerListeners,
             activityOrFragment = activityOrFragment,
             parent = parent
         )
@@ -201,7 +192,7 @@ fun registerElement(
         )
     }
     if (registerListeners) {
-        registerListeners(view)
+        registerListeners(view, logger)
     }
 }
 
@@ -349,206 +340,4 @@ fun registerComponent(
                 rts = rts
             )
         )
-}
-
-// list of text watchers in the entire app
-val textWatchers = mutableListOf<TextWatcher>()
-
-private fun registerListeners(view: View) {
-    val idName = view.getIdOrTag()
-    val simpleClassName = view.javaClass.simpleName
-    val gyroData = NIDSensorHelper.getGyroscopeInfo()
-    val accelData = NIDSensorHelper.getAccelerometerInfo()
-
-    // EditText is a parent class to multiple components
-    if (view is EditText) {
-        NIDLog.d(
-            "NID-Activity",
-            "EditText Listener $simpleClassName - ${view::class} - ${view.getIdOrTag()}"
-        )
-        // add Text Change watcher
-        val textWatcher = NIDTextWatcher(
-            idName,
-            simpleClassName,
-            "${view.text.toString().getSHA256withSalt()?.take(8)}"
-        )
-        // first we have to clear the text watcher that is currently in the EditText
-        for (watcher in textWatchers) {
-            view.removeTextChangedListener(watcher)
-        }
-        // we add the new one in there
-        view.addTextChangedListener(textWatcher)
-        // we add the new one to the list of existing text watchers so we can remove it later when
-        // it is re-registered
-        textWatchers.add(textWatcher)
-
-        // add original action menu watcher
-        val actionCallback = view.customSelectionActionModeCallback
-        if (actionCallback !is NIDTextContextMenuCallbacks) {
-            view.customSelectionActionModeCallback = NIDTextContextMenuCallbacks(actionCallback)
-        }
-
-        // if later api version, add additional action menu watcher
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            addExtraActionMenuListener(view)
-        }
-    }
-
-    when (view) {
-        is AbsSpinner -> {
-            val lastClickListener = view.onItemClickListener
-            view.onItemClickListener = null
-            view.onItemClickListener =
-                addSelectOnClickListener(
-                    idName,
-                    lastClickListener,
-                    simpleClassName,
-                    gyroData,
-                    accelData
-                )
-
-            val lastSelectListener = view.onItemSelectedListener
-            view.onItemSelectedListener = null
-            view.onItemSelectedListener = addSelectOnSelect(
-                idName,
-                lastSelectListener,
-                simpleClassName,
-                gyroData,
-                accelData
-            )
-        }
-
-        is Spinner -> {
-            val lastClickListener = view.onItemClickListener
-            view.onItemClickListener = null
-            view.onItemClickListener =
-                addSelectOnClickListener(
-                    idName,
-                    lastClickListener,
-                    simpleClassName,
-                    gyroData,
-                    accelData
-                )
-
-            val lastSelectListener = view.onItemSelectedListener
-            view.onItemSelectedListener = null
-            view.onItemSelectedListener = addSelectOnSelect(
-                idName,
-                lastSelectListener,
-                simpleClassName,
-                gyroData,
-                accelData
-            )
-        }
-
-        is AutoCompleteTextView -> {
-            val lastClickListener = view.onItemClickListener
-            view.onItemClickListener = null
-            view.onItemClickListener =
-                addSelectOnClickListener(
-                    idName,
-                    lastClickListener,
-                    simpleClassName,
-                    gyroData,
-                    accelData
-                )
-
-            val lastSelectListener = view.onItemSelectedListener
-            view.onItemSelectedListener = null
-            view.onItemSelectedListener = addSelectOnSelect(
-                idName,
-                lastSelectListener,
-                simpleClassName,
-                gyroData,
-                accelData
-            )
-        }
-    }
-}
-
-
-private fun addSelectOnSelect(
-    idName: String,
-    lastSelectListener: AdapterView.OnItemSelectedListener?,
-    simpleClassName: String,
-    gyroData: NIDSensorModel?,
-    accelData: NIDSensorModel?,
-): AdapterView.OnItemSelectedListener {
-    return object : AdapterView.OnItemSelectedListener {
-        override fun onItemSelected(
-            adapter: AdapterView<*>?,
-            viewList: View?,
-            position: Int,
-            p3: Long
-        ) {
-            lastSelectListener?.onItemSelected(adapter, viewList, position, p3)
-
-//            NIDLog.d(
-//                "NID-Activity",
-//                "Select Selected $idName - $simpleClassName $position $p3 $viewList"
-//            )
-            getDataStoreInstance()
-                .saveEvent(
-                    NIDEventModel(
-                        type = SELECT_CHANGE,
-                        tg = hashMapOf(
-                            "etn" to simpleClassName,
-                            "tgs" to idName,
-                            "sender" to simpleClassName
-                        ),
-                        tgs = idName,
-                        ts = System.currentTimeMillis(),
-                        gyro = gyroData,
-                        accel = accelData,
-                        v = "$position"
-                    )
-                )
-        }
-
-        override fun onNothingSelected(p0: AdapterView<*>?) {
-            lastSelectListener?.onNothingSelected(p0)
-        }
-    }
-}
-
-private fun addSelectOnClickListener(
-    idName: String,
-    lastClickListener: AdapterView.OnItemClickListener?,
-    simpleClassName: String,
-    gyroData: NIDSensorModel?,
-    accelData: NIDSensorModel?,
-): AdapterView.OnItemClickListener {
-    return AdapterView.OnItemClickListener { adapter, viewList, position, p3 ->
-        lastClickListener?.onItemClick(adapter, viewList, position, p3)
-
-//        NIDLog.d(
-//            "NID-Activity",
-//            "Select CLICK $idName - $simpleClassName $position $p3 $viewList"
-//        )
-        getDataStoreInstance()
-            .saveEvent(
-                NIDEventModel(
-                    type = SELECT_CHANGE,
-                    tg = hashMapOf(
-                        "etn" to "INPUT",
-                        "et" to "text"
-                    ),
-                    tgs = idName,
-                    ts = System.currentTimeMillis(),
-                    gyro = gyroData,
-                    accel = accelData,
-                    v = "$position"
-                )
-            )
-    }
-}
-
-
-@RequiresApi(Build.VERSION_CODES.M)
-private fun addExtraActionMenuListener(view: EditText) {
-    val actionInsertionCallback = view.customInsertionActionModeCallback
-    if (actionInsertionCallback !is NIDLongPressContextMenuCallbacks) {
-        view.customInsertionActionModeCallback =
-            NIDLongPressContextMenuCallbacks(actionInsertionCallback)
-    }
 }
