+++ NeuroID/src/reactNativeLib/java/com/neuroid/tracker/events/NIDRegisterEventsHelper.kt
@@ -4,21 +4,10 @@ import android.app.Activity
 import android.os.Looper
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity
 import com.neuroid.tracker.callbacks.*
-import com.neuroid.tracker.service.NIDServiceTracker
 import com.neuroid.tracker.storage.NIDDataStoreManager
 import com.neuroid.tracker.utils.NIDLogWrapper
 import java.util.*
-
-fun registerLaterLifecycleFragments(activity: Activity) {
-    NIDServiceTracker.screenActivityName = activity::class.java.name
-
-    val fragManager = (activity as? AppCompatActivity)?.supportFragmentManager
-    fragManager?.registerFragmentLifecycleCallbacks(NIDFragmentCallbacks(), true)
-
-    registerWindowListeners(activity)
-}
 
 fun registerWindowListeners(activity: Activity) {
     val viewMainContainer = activity.window.decorView.findViewById<View>(
@@ -62,12 +51,13 @@ fun registerTargetFromScreen(
         identifyAllViews(
             viewMainContainer,
             guid,
-            registerTarget,
-            registerListeners,
             logger,
             storeManager,
+            registerTarget,
+            registerListeners,
             activityOrFragment,
             parent
         )
     }, 300)
 }
+
